+++ tests/testthat/test-col_types.R
@@ -10,3 +10,10 @@ test_that("You can use !! in cols and cols_only", {
     cols_only(xyz = col_character())
   )
 })
+
+test_that("format(col_spec) contains the delimiter if specified", {
+  expect_match(fixed = TRUE,
+    format(cols(.delim = "\t")),
+    '.delim = "\\t"'
+  )
+})